Updated zuix.js dist/CLI.

+++ .eleventy-zuix.js
@@ -41,6 +41,7 @@ const {classNameFromHyphens} = require('zuix/common/utils');
 const chalk = require('chalk');
 const mkdirp = require('mkdirp');
 const yesno = require('yesno');
+const capcon = require('capture-console');
 
 // Read configuration either from './config/{default}.json'
 // or './config/production.json' based on current `NODE_ENV'
@@ -56,9 +57,6 @@ const contentSourceFolder = path.join(sourceFolder, contentFolder);
 const contentBuildFolder = path.join(buildFolder, contentFolder)
 const dataFolder = zuixConfig.get('build.dataFolder');
 const includesFolder = zuixConfig.get('build.includesFolder');
-// this file is a temporary file create to trigger 11ty build
-const triggerFile = path.join(sourceFolder, '.zuix.build.md');
-const triggerFileOut = path.join(buildFolder, '.zuix.build.tmp');
 // replace {{variables}} eventually employed in the config
 zuixConfig = JSON.parse(nunjucks.renderString(JSON.stringify(zuixConfig), zuixConfig));
 
@@ -97,16 +95,40 @@ function startWatcher(eleventyConfig, browserSync) {
     f = path.resolve(path.join(sourceFolder, f));
     watchFiles.push(f);
   });
+  const templateExtensions = ['.html', '.js', '.css', '.less', '.scss', '.njk'];
+  const templateFolders = componentsFolders.map(f =>  path.resolve(path.join(sourceFolder, f)));
   const copyFilesWatcher = chokidar.watch(watchFiles).on('all', (event, file, stats) => {
-    if (watchEvents[event] && fs.existsSync(file)) {
+    if (watchEvents[event] && fs.existsSync(file) && file.indexOf('/_inc/') === -1) {
       const outputFile = path.resolve(path.join(buildFolder, file.substring(path.resolve(sourceFolder).length)));
       const outputFolder = path.dirname(outputFile);
       if (!fs.existsSync(outputFolder)) {
         fs.mkdirSync(outputFolder, { recursive: true })
       }
-      fs.copyFileSync(file, outputFile);
+      const postProcess = templateExtensions.filter(cn => file.endsWith(cn));
+      if (postProcess.length === 1) {
+        // Post-process file with Nunjucks
+        const njk = new nunjucks.Environment(new nunjucks.FileSystemLoader([
+          path.dirname(file),
+          ...templateFolders
+        ], {}));
+        njk.render(file, zuixConfig, function(err, res) {
+          if (err != null) {
+            console.error(
+              chalk.red.bold(err)
+            );
+          } else {
+            fs.writeFile(outputFile, res, function() {
+              // TODO: ...
+            });
+          }
+        });
+      } else {
+        // Do not post-process, copy as-is
+        fs.copyFileSync(file, outputFile);
+      }
     } else {
-      // TODO: maybe remove file from output folder as well?
+      // TODO: maybe remove file from output folder as well if it was unlinked?
+      // TODO: >> recompile dependant files if the modified file is inside an `_inc` folder
     }
     if (browserSync) {
       browserSync.reload();
@@ -114,67 +136,76 @@ function startWatcher(eleventyConfig, browserSync) {
   });
   const includes = path.join(sourceFolder, includesFolder);
   const allFilesWatcher = chokidar.watch([sourceFolder]).on('all', (event, file, stats) => {
-    if (event.startsWith('unlink') && file !== triggerFile) {
-      forceRebuild();
+    if (event.startsWith('unlink')) {
+      forceRebuild(200);
     } else if (!file.startsWith(includes) && file.indexOf(path.join('/', '_inc', '/')) !== -1) {
-      forceRebuild();
+      forceRebuild(200);
     }
   });
   setupSocketApi(browserSync);
 }
 
 function setupSocketApi(browserSync) {
-  // WebSocket API
+  // zuix-browser-sync WebSocket API
   if (browserSync && !io) {
     io = browserSync.instance.io;
     io.on('connection', (socket) => {
       socket.on('zuix:loadContent', (request) => {
-        const content = fs.readFileSync(request.path);
-        request.content = content.toString('utf8');
-        io.emit('zuix:loadContent:done', request);
-        console.log('zuix:loadContent', request.path);
+        cmsLog('zuix:loadContent', request.path);
+        fs.readFile(request.path, (err, content) => {
+          if (!err) {
+            request.content = content.toString('utf8');
+            ioEmit('zuix:loadContent:done', request);
+          } else {
+            ioEmit('zuix:loadContent:error', err.message);
+          }
+        });
       });
       socket.on('zuix:saveContent', (request) => {
-        fs.writeFileSync(request.path, request.content);
-        io.emit('zuix:saveContent:done', {path: request.path});
-        console.log('zuix:saveContent', request.path);
+        cmsLog('zuix:saveContent', request.path);
+        fs.writeFile(request.path, request.content, () => {
+          ioEmit('zuix:saveContent:done', {path: request.path});
+          forceRebuild(2000); // <--- TODO: this is a patch to the fact sometimes 11ty doesn't rebuild
+        });
       });
       socket.on('zuix:addPage', (data) => {
-        browserSync.notify('Adding new page...');
+        cmsLog('zuix:addPage', data);
         addPage(data).then(() => {
           const redirectUrl = zuixConfig.app.baseUrl + contentFolder + '/' + data.section + '/';
-          io.emit('zuix:addPage:done', redirectUrl);
+          ioEmit('zuix:addPage:done', redirectUrl);
+          forceRebuild(2000); // <--- TODO: this is a patch to the fact sometimes 11ty doesn't rebuild
         }).catch((err) => {
-          io.emit('zuix:addPage:error', err.message);
+          ioEmit('zuix:addPage:error', err.message);
         });
-        console.log('zuix:addPage', data);
       });
       socket.on('zuix:deletePage', (data) => {
-        browserSync.notify('Deleting page...');
+        cmsLog('zuix:deletePage', data.page.outputPath, path.resolve(data.page.url, '..'));
         try {
           if (data.page.filePathStem.endsWith('/index')) {
             // delete folder too
             fs.rmSync(path.resolve(data.page.inputPath, '..'), { recursive: true, force: true });
             fs.rmSync(path.resolve(data.page.outputPath, '..'), { recursive: true, force: true });
+          } else {
+            // delete single file
+            fs.rmSync(data.page.inputPath);
           }
           const redirectUrl = path.resolve(data.page.url, '..');
-          io.emit('zuix:deletePage:done', redirectUrl);
+          ioEmit('zuix:deletePage:done', redirectUrl);
+          forceRebuild();
         } catch (e) {
-          console.log(e);
-          io.emit('zuix:deletePage:error', e);
+          ioEmit('zuix:deletePage:error', e);
         }
-        console.log('zuix:deletePage', data.page.outputPath, path.resolve(data.page.url, '..'));
       });
       socket.on('zuix:addComponent', (data) => {
         browserSync.notify('Adding component...');
         const handlePromise = (promise) => {
           promise.then((res) =>  {
             generateTemplate(res);
-            io.emit('zuix:addComponent:done', res);
+            ioEmit('zuix:addComponent:done', res);
           })
-          .catch((err) => {
-            io.emit('zuix:addComponent:error', err.message);
-          });
+            .catch((err) => {
+              ioEmit('zuix:addComponent:error', err.message);
+            });
         };
         if (data.view && data.ctrl) {
           handlePromise(generate('component', [data.name]));
@@ -184,15 +215,77 @@ function setupSocketApi(browserSync) {
           handlePromise(generate('controller', [data.name]));
         }
       });
+      socket.on('zuix:buildAll', (request) => {
+        cmsLog('zuix:buildAll');
+        forceRebuild(200);
+      });
+    });
+    // Errors notifying
+    let errorObject = {
+      errors: [],
+      message: '',
+      debug: false
+    }
+    // Eleventy console messages handling
+    let errorNotifierTimeout = null;
+    const errorNotifier = function() {
+      if (!errorObject.errors[0].startsWith('! ') && !errorObject.errors[0].startsWith(' Benchmark ')) {
+        ioEmit('zuix:eleventy:error', errorObject);
+      }
+      errorObject.errors = [];
+      errorObject.message = '';
+      errorObject.debug = false;
+    }
+    // Capture console output to notify errors
+    capcon.startCapture(process.stderr, {}, function (stderr) {
+      if (stderr && stderr.replace) {
+        stderr = stderr
+            .replace(/[\u001b\u009b][[()#;?]*(?:[0-9]{1,4}(?:;[0-9]{0,4})*)?[0-9A-ORZcf-nqry=><]/g, '')
+            .replace(/ *\([^)]*\) */g, '').replace(/ +$/, '');
+        if (stderr.indexOf('\n[11ty] ') !== -1) {
+          errorObject.debug = true;
+        } else if (stderr.startsWith('[11ty]')) {
+          stderr = stderr.substring(6).replace(/ +$/, '');
+        }
+        errorObject.errors.push(stderr);
+        if (!errorObject.debug) {
+          errorObject.message += stderr;
+        }
+        clearTimeout(errorNotifierTimeout);
+        errorNotifierTimeout = setTimeout(errorNotifier, 250);
+      }
     });
   }
 }
+function ioEmit(message, data) {
+  io.emit(message, data);
+  // log emitted data
+  let info = data ? ` (${JSON.stringify(data)})` : '';
+  if (data && data.page) {
+    info = ` (${data.page.outputPath} ${path.resolve(data.page.url, '..')})`;
+  } else if (data && data.path) {
+    info = ` (${data.path})`;
+  }
+  cmsLog(`${message}${info}`);
+}
+function cmsLog(...args) {
+  console.log(`[zuix-cms] ${args.join(' ')}`);
+}
 
-function forceRebuild() {
-  fs.writeFileSync(triggerFile, `---
-permalink: .zuix.build.tmp
----
-Temporary file to trigger 11ty build`);
+let forceRebuildTimeout = null;
+function forceRebuild(delay) {
+  clearTimeout(forceRebuildTimeout);
+  forceRebuildTimeout = setTimeout(() => {
+    touch('.eleventy.js');
+  }, delay);
+}
+function touch(filename) {
+  const time = new Date();
+  try {
+    fs.utimesSync(filename, time, time);
+  } catch (err) {
+    fs.closeSync(fs.openSync(filename, 'w'));
+  }
 }
 
 function initEleventyZuix(eleventyConfig) {
@@ -226,6 +319,7 @@ function initEleventyZuix(eleventyConfig) {
     return content;
   });
   eleventyConfig.on('eleventy.beforeWatch', (cf) => {
+    clearTimeout(forceRebuildTimeout);
     // changedFiles is an array of files that changed
     // to trigger the watch/serve build
     changedFiles.length = 0;
@@ -263,13 +357,8 @@ function initEleventyZuix(eleventyConfig) {
       // reverts to incremental build mode
       rebuildAll = false;
     }
-    // delete temporary build-trigger file if found
-    if (fs.existsSync(triggerFile)) {
-      fs.unlinkSync(triggerFile);
-      fs.unlinkSync(triggerFileOut);
-    }
     if (io) {
-      io.emit('zuix:build:done');
+      ioEmit('zuix:build:done');
     }
   });
   if (process.argv.indexOf('--serve') === -1) {
@@ -335,7 +424,6 @@ function generateTemplate(data) {
   if (!fs.existsSync(outputFile)) {
     mkdirp.sync(path.dirname(outputFile));
     fs.writeFileSync(outputFile, `const template = '${data.html}';
-
 module.exports = (render, content, linkUrl) => {
   return render(template, {content, linkUrl});
 };
@@ -472,22 +560,24 @@ function addPage(args) {
       outputFile = path.join(outputPath, 'index' + extension);
       if (!fs.existsSync(outputFile)) {
         mkdirp.sync(outputPath);
-        fs.writeFileSync(outputFile, pageTemplate);
-        console.log(chalk.cyanBright('*') + ' NEW page:', chalk.green.bold(outputFile));
-        // Create section if it does not exist
-        if (args.section) {
-          const sectionFile = path.join(sectionFolder, 'index.liquid');
-          if (!fs.existsSync(sectionFile)) {
-            addPage({layout: 'section', name: args.section, frontMatter: [
-                `group: ${args.section}`,
-                `title: ${classNameFromHyphens(args.section)}`,
-              ]}).then(resolve).catch(reject);
+        fs.writeFile(outputFile, pageTemplate, () => {
+          console.log(chalk.cyanBright('*') + ' NEW page:', chalk.green.bold(outputFile));
+          // Create section if it does not exist
+          if (args.section) {
+            const sectionFile = path.join(sectionFolder, 'index.liquid');
+            if (!fs.existsSync(sectionFile)) {
+              addPage({layout: 'section', name: args.section, frontMatter: [
+                  `group: ${args.section}`,
+                  `title: ${classNameFromHyphens(args.section)}`,
+                ]}).then(resolve).catch(reject);
+            } else {
+//              forceRebuild(3000);
+              resolve();
+            }
           } else {
             resolve();
           }
-        } else {
-          resolve();
-        }
+        });
       } else {
         const errorMessage = `"${args.name}" already exists.`;
         console.error(
